Adicionado CRUD completo de sistema, setor e cargo. começando a fazer relacionamentos

+++ src/controllers/pessoasController.js
@@ -8,10 +8,10 @@ class PessoasController {
   static async buscaTodasPessoas (req, res, next) {
     const pessoasEncontradas = await database.TN_T_PESSOA.findAll();
     try{
-      if(pessoasEncontradas !== null){
+      if(pessoasEncontradas.length !== 0){
         res.status(200).send(pessoasEncontradas);
       } else {
-        next(new NaoEncontrado());
+        next(new NaoEncontrado("Nenhuma pessoa cadastrada."));
       }
     } catch (err) {
       next(err);
@@ -20,21 +20,17 @@ class PessoasController {
 
   static async buscaPessoaPorId (req, res, next) {
     const { id } = req.params;
-    const pessoaEncontrada = await database.TN_T_PESSOA.findOne({
-      where: { 
-        id: Number(id) 
-      }});
+    const pessoaEncontrada = await database.TN_T_PESSOA.findOne({ where: {id: Number(id)}});
     try{
-      if(pessoaEncontrada !== null){
+      if(pessoaEncontrada){
         res.status(200).send(pessoaEncontrada);
       } else {
-        next(new NaoEncontrado());
+        next(new NaoEncontrado(`ID ${id} de pessoa não encontrada na busca.`));
       }
     } catch (err) {
       next(err);
     }
   }
-
 
   static async criaUmaPessoa(req, res, next) {
     const novaPessoa = {
@@ -58,11 +54,24 @@ class PessoasController {
 
   static async atualizaUmaPessoa(req, res, next) {
     const { id } = req.params;
-    const novasInfos = req.body;
+    const pessoaEncontrada = await database.TN_T_PESSOA.findOne({ where: {id: Number(id)}});
+    const novasInfos = {
+      ds_nome: req.body.ds_nome,
+      nr_cpf: req.body.nr_cpf,
+      dt_admissao: req.body.dt_admissao,
+      dt_nascimento: req.body.dt_nascimento,
+      tp_contrato: req.body.tp_contrato,
+      ds_categoria_cargo: req.body.tp_categoria_cargo,
+      dt_updated: new Date(),
+    };
     try{
-      await database.TN_T_PESSOA.update(novasInfos, {where: {id: Number(id)}});
-      const pessoaAtualizada = await database.TN_T_PESSOA.findOne( { where:{ id: Number(id) }});
-      res.status(200).send(pessoaAtualizada);
+      if(pessoaEncontrada){
+        await database.TN_T_PESSOA.update(novasInfos, {where: {id: Number(id)}});
+        const pessoaAtualizada = await database.TN_T_PESSOA.findOne( { where:{ id: Number(id) }});
+        res.status(200).send(pessoaAtualizada);
+      } else {
+        next(new NaoEncontrado(`ID ${id} de pessoa não encontrada para atualizar informações.`));
+      }
     } catch (err) {
       next(err);
     }
@@ -70,9 +79,24 @@ class PessoasController {
 
   static async deletaUmaPessoa(req, res, next) {
     const {id} = req.params;
+    const pessoaEncontrada = await database.TN_T_PESSOA.findOne({ where: {id: Number(id)}});
     try{
-      await database.TN_T_PESSOA.destroy({where: {id: Number(id)}});
-      res.status(200).send({message: `Usuário de ID ${id} excluído`});
+      if(pessoaEncontrada){
+        await database.TN_T_PESSOA.destroy({where: {id: Number(id)}});
+        res.status(200).send({message: `Usuário de ID ${id} excluído.`});
+      } else {
+        next(new NaoEncontrado(`ID ${id} de pessoa não encontrada para excluir.`));
+      }
+    } catch (err) {
+      next(err);
+    }
+  }
+
+  static async restauraUmaPessoa(req, res, next) {
+    const { id } = req.params;
+    try{
+      await database.TN_T_PESSOA.restore({where : {id : Number(id)}});
+      res.status(200).send({message: `ID ${id} restaurado.`});
     } catch (err) {
       next(err);
     }
